MSI-1431: refactoring
