+++ app/code/Magento/InventorySales/composer.json
@@ -11,7 +11,7 @@
         "magento/module-inventory-configuration-api": "*",
         "magento/module-inventory-reservations-api": "*",
         "magento/module-inventory-sales-api": "*",
-        "magento/module-inventory-source-deduction": "*",
+        "magento/module-inventory-source-deduction-api": "*",
         "magento/module-sales-inventory": "*",
         "magento/module-store": "*",
         "magento/module-sales": "*"
