+++ app/code/Magento/InventoryShipping/Observer/VirtualSourceDeductionProcessor.php
@@ -13,8 +13,8 @@ use Magento\Framework\Exception\LocalizedException;
 use Magento\InventorySalesApi\Api\Data\SalesEventInterface;
 use Magento\InventorySalesApi\Api\Data\SalesEventInterfaceFactory;
 use Magento\InventoryShipping\Model\GetSourceSelectionResultFromInvoice;
-use Magento\InventorySourceDeduction\Model\SourceDeductionRequestInterface;
-use Magento\InventorySourceDeduction\Model\SourceDeductionServiceInterface;
+use Magento\InventorySourceDeductionApi\Model\SourceDeductionRequestInterface;
+use Magento\InventorySourceDeductionApi\Model\SourceDeductionServiceInterface;
 use Magento\InventoryShipping\Model\SourceDeductionRequestsFromSourceSelectionFactory;
 use Magento\Sales\Api\Data\InvoiceInterface;
 use Magento\Sales\Api\Data\OrderItemInterface;
