--- app/code/Magento/InventorySourceDeduction/etc/di.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\InventorySourceDeduction\Model\ItemToDeductInterface" type="Magento\InventorySourceDeduction\Model\ItemToDeduct"/>
-    <preference for="Magento\InventorySourceDeduction\Model\SourceDeductionRequestInterface" type="Magento\InventorySourceDeduction\Model\SourceDeductionRequest"/>
-    <preference for="Magento\InventorySourceDeduction\Model\SourceDeductionServiceInterface" type="Magento\InventorySourceDeduction\Model\SourceDeductionService"/>
-</config>
