+++ app/code/Magento/InventorySourceDeductionApi/LICENSE.txt
