+++ app/code/Magento/InventorySourceDeductionApi/LICENSE_AFL.txt
