+++ app/code/Magento/InventorySourceDeductionApi/Model/GetSourceItemBySourceCodeAndSku.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventorySourceDeduction\Model;
+namespace Magento\InventorySourceDeductionApi\Model;
 
 use Magento\Framework\Api\SearchCriteriaBuilder;
 use Magento\InventoryApi\Api\Data\SourceItemInterface;
