+++ app/code/Magento/InventorySourceDeductionApi/Model/SourceDeductionService.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventorySourceDeduction\Model;
+namespace Magento\InventorySourceDeductionApi\Model;
 
 use Magento\InventoryApi\Api\SourceItemsSaveInterface;
 use Magento\InventoryConfigurationApi\Api\GetStockItemConfigurationInterface;
