+++ app/code/Magento/InventorySourceDeductionApi/composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "magento/module-inventory-source-deduction",
+  "name": "magento/module-inventory-source-deduction-api",
   "description": "N/A",
   "require": {
     "php": "~7.1.3||~7.2.0",
@@ -18,7 +18,7 @@
       "registration.php"
     ],
     "psr-4": {
-      "Magento\\InventorySourceDeduction\\": ""
+      "Magento\\InventorySourceDeductionApi\\": ""
     }
   }
 }
