+++ app/code/Magento/InventorySourceDeductionApi/etc/di.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <preference for="Magento\InventorySourceDeductionApi\Model\ItemToDeductInterface" type="Magento\InventorySourceDeductionApi\Model\ItemToDeduct"/>
+    <preference for="Magento\InventorySourceDeductionApi\Model\SourceDeductionRequestInterface" type="Magento\InventorySourceDeductionApi\Model\SourceDeductionRequest"/>
+    <preference for="Magento\InventorySourceDeductionApi\Model\SourceDeductionServiceInterface" type="Magento\InventorySourceDeductionApi\Model\SourceDeductionService"/>
+</config>
