+++ app/code/Magento/InventorySourceDeductionApi/etc/module.xml
@@ -6,5 +6,5 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_InventorySourceDeduction" setup_version="1.0.0" />
+    <module name="Magento_InventorySourceDeductionApi" setup_version="1.0.0" />
 </config>
