+++ app/code/Magento/InventorySourceDeductionApi/registration.php
@@ -7,6 +7,6 @@ declare(strict_types=1);
 
 \Magento\Framework\Component\ComponentRegistrar::register(
     \Magento\Framework\Component\ComponentRegistrar::MODULE,
-    'Magento_InventorySourceDeduction',
+    'Magento_InventorySourceDeductionApi',
     __DIR__
 );
