+++ composer.json
@@ -195,7 +195,7 @@
         "magento/module-inventory-sales-frontend-ui": "*",
         "magento/module-inventory-shipping": "*",
         "magento/module-inventory-shipping-admin-ui": "*",
-        "magento/module-inventory-source-deduction": "*",
+        "magento/module-inventory-source-deduction-api": "*",
         "magento/module-inventory-source-selection-api": "*",
         "magento/module-inventory-source-selection": "*",
         "magento/module-layered-navigation": "*",
